+++ modules/SofaBaseVisual/VisualModelImpl.cpp
@@ -289,7 +289,7 @@ void VisualModelImpl::setMesh(helper::io::Mesh &objLoader, bool tex)
             else if (verts.size() == 4)
                 nbf[NBQ] += 1;
             else
-                nbf[NBT] += verts.size()-2;
+                nbf[NBT] += (int)verts.size()-2;
         }
         facet2tq[facetsImport.size()] = nbf;
         groups.resize(objLoader.getGroups().size());
@@ -313,7 +313,7 @@ void VisualModelImpl::setMesh(helper::io::Mesh &objLoader, bool tex)
         }
     }
 
-    int nbVIn = verticesImport.size();
+    size_t nbVIn = verticesImport.size();
     // First we compute for each point how many pair of normal/texcoord indices are used
     // The map store the final index of each combinaison
     vector< std::map< std::pair<int,int>, int > > vertTexNormMap;
@@ -332,11 +332,11 @@ void VisualModelImpl::setMesh(helper::io::Mesh &objLoader, bool tex)
     }
 
     // Then we can compute how many vertices are created
-    int nbVOut = 0;
+    size_t nbVOut = 0;
     bool vsplit = false;
-    for (int i = 0; i < nbVIn; i++)
+    for (size_t i = 0; i < nbVIn; i++)
     {
-        int s = vertTexNormMap[i].size();
+        size_t s = vertTexNormMap[i].size();
         nbVOut += s;
     }
 
@@ -410,8 +410,8 @@ void VisualModelImpl::setMesh(helper::io::Mesh &objLoader, bool tex)
     }
 
     if (!vsplit)
-        nbNOut = nbVOut;
-    else if (nbNOut == nbVOut)
+        nbNOut = (int)nbVOut;
+    else if (nbNOut == (int)nbVOut)
         vertNormIdx.resize(0);
 
 
@@ -818,13 +818,13 @@ void VisualModelImpl::computeNormals()
 
     if (vertNormIdx.empty())
     {
-        int nbn = (vertices).size();
+        size_t nbn = (vertices).size();
         //serr << "CN0("<<nbn<<")"<<sendl;
 
         ResizableExtVector<Deriv>& normals = *(m_vnormals.beginEdit());
 
         normals.resize(nbn);
-        for (int i = 0; i < nbn; i++)
+        for (size_t i = 0; i < nbn; i++)
             normals[i].clear();
 
         for (unsigned int i = 0; i < triangles.size(); i++)
@@ -1419,7 +1419,7 @@ void VisualModelImpl::handleTopologyChange()
             const sofa::core::topology::TrianglesAdded *ta = static_cast< const sofa::core::topology::TrianglesAdded * >( *itBegin );
             Triangle t;
             const unsigned int nbAddedTriangles = ta->getNbAddedTriangles();
-            const unsigned int nbTririangles = triangles.size();
+            const size_t nbTririangles = triangles.size();
             triangles.resize(nbTririangles + nbAddedTriangles);
 
             for (unsigned int i = 0; i < nbAddedTriangles; ++i)
@@ -1444,7 +1444,7 @@ void VisualModelImpl::handleTopologyChange()
             const sofa::core::topology::QuadsAdded *qa = static_cast< const sofa::core::topology::QuadsAdded * >( *itBegin );
             Quad q;
             const unsigned int nbAddedQuads = qa->getNbAddedQuads();
-            const unsigned int nbQuaduads = quads.size();
+            const size_t nbQuaduads = quads.size();
             quads.resize(nbQuaduads + nbAddedQuads);
 
             for (unsigned int i = 0; i < nbAddedQuads; ++i)
@@ -1484,7 +1484,7 @@ void VisualModelImpl::handleTopologyChange()
                 triangles[ind_k] = triangles[last];
                 triangles[last] = tmp;
 
-                unsigned int ind_last = triangles.size() - 1;
+                size_t ind_last = triangles.size() - 1;
 
                 if(last != ind_last)
                 {
@@ -1525,7 +1525,7 @@ void VisualModelImpl::handleTopologyChange()
                 quads[ind_k] = quads[last];
                 quads[last] = tmp;
 
-                unsigned int ind_last = quads.size() - 1;
+                size_t  ind_last = quads.size() - 1;
 
                 if(last != ind_last)
                 {
@@ -1856,7 +1856,7 @@ void VisualModelImpl::exportOBJ(std::string name, std::ostream* out, std::ostrea
     const ResizableExtVector<int> &vertPosIdx = m_vertPosIdx.getValue();
     const ResizableExtVector<int> &vertNormIdx = m_vertNormIdx.getValue();
 
-    int nbv = x.size();
+    int nbv = (int)x.size();
 
     for (unsigned int i=0; i<x.size(); i++)
     {
@@ -1867,7 +1867,7 @@ void VisualModelImpl::exportOBJ(std::string name, std::ostream* out, std::ostrea
 
     if (vertNormIdx.empty())
     {
-        nbn = vnormals.size();
+        nbn = (int)vnormals.size();
         for (unsigned int i=0; i<vnormals.size(); i++)
         {
             *out << "vn "<< std::fixed << vnormals[i][0]<<' '<< std::fixed <<vnormals[i][1]<<' '<< std::fixed <<vnormals[i][2]<<'\n';
@@ -1895,7 +1895,7 @@ void VisualModelImpl::exportOBJ(std::string name, std::ostream* out, std::ostrea
     int nbt = 0;
     if (!vtexcoords.empty())
     {
-        nbt = vtexcoords.size();
+        nbt = (int)vtexcoords.size();
         for (unsigned int i=0; i<vtexcoords.size(); i++)
         {
             *out << "vt "<< std::fixed << vtexcoords[i][0]<<' '<< std::fixed <<vtexcoords[i][1]<<'\n';
